remove warning
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@7097 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ dune/fem/space/common/cachedcommmanager.hh
@@ -506,7 +506,7 @@ namespace Dune
 
       const int size = indexMap.size();
       // make sure that the receive buffer has the correct size 
-      assert( size * blockSize * sizeof( DofType ) <= str.size() );
+      assert( size_t(size * blockSize * sizeof( DofType )) <= size_t(str.size()) );
       for( int i = 0; i < size; ++i )
       {
         // get dof block
